Move dwave::optimization::functional namespace into its own file
