+++ tests/cpp/meson.build
@@ -10,7 +10,6 @@ tests_all = executable(
         'nodes/indexing/test_advanced.cpp',
         'nodes/indexing/test_basic.cpp',
         'nodes/mathematical/test_binaryop.cpp',
-        'nodes/mathematical/test_functional.cpp',
         'nodes/mathematical/test_softmax.cpp',
         'nodes/mathematical/test_naryop.cpp',
         'nodes/mathematical/test_reduce.cpp',
@@ -29,6 +28,7 @@ tests_all = executable(
         'nodes/test_sorting.cpp',
         'nodes/test_statistics.cpp',
         'test_array.cpp',
+        'test_functional.cpp',
         'test_graph.cpp',
         'test_iterators.cpp',
         'test_simplex.cpp',
